+++ src/services/interFace.ts
@@ -1,0 +1,9 @@
+export  interface userData {
+    userName: string,
+    email: string,
+    country: string,
+    phoneNumber: number,
+    password: string | number,
+    confirmPassword: string | number
+
+  }
